+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Revision history for tasty-test
 
+## Unreleased
+
+* Add commandline option "--remove-tix-hash" which replaces the hash in the tix files with "0". This is useful for golden testing the output of the tasty-coverage ingredient.
+ 
 ## 0.1.0.0 -- 2023-08-07
 
 * Release of first version on Hackage