+++ CHANGELOG.md
@@ -1,11 +1,11 @@
 # Revision history for tasty-test
 
-## Unreleased
+## 0.1.1.0 -- 2023-09-07
 
 * Be more specific with the file suffixes which are computed from the test outcome. Previously only "PASSED" and "FAILED" were used. We now also use "EXCEPTION", "TIMEOUT" and "SKIPPED".
 * If the name of the test contains a path separator, then this path separator is removed from the name of the generated tix file (#5)
 * Add commandline option "--remove-tix-hash" which replaces the hash in the tix files with "0". This is useful for golden testing the output of the tasty-coverage ingredient.
 
-## 0.1.0.0 -- 2023-08-07
+## 0.1.0.0 -- 2023-08-21
 
 * Release of first version on Hackage